change diddyborg to thunderborg
