+++ controller.sample.conf
@@ -280,7 +280,7 @@ drivingSpeed=180
 [max7219]
 ledrotate=180
 
-[diddyborg]
+[thunderborg]
 # maximum is 1.0, minimum is 0.0
 left_motor_max=1.0
 right_motor_max=1.0
